Moving AssertingEventsClient to the prefect.testing package.  Fixing __aexit__s
