+++ src/prefect/events/clients.py
@@ -1,8 +1,8 @@
 import abc
 from types import TracebackType
-from typing import List, Optional, Type
+from typing import Optional, Type
 
-from .schemas import Event
+from prefect.events.schemas import Event
 
 
 class EventsClient(abc.ABC):
@@ -20,8 +20,8 @@ class EventsClient(abc.ABC):
         exc_type: Optional[Type[Exception]],
         exc_val: Optional[Exception],
         exc_tb: Optional[TracebackType],
-    ) -> "EventsClient":
-        return self
+    ) -> Optional[bool]:
+        return None
 
 
 class NullEventsClient(EventsClient):
@@ -29,29 +29,3 @@ class NullEventsClient(EventsClient):
 
     async def emit(self, event: Event) -> None:
         pass
-
-
-class AssertingEventsClient(EventsClient):
-    """An implementation of the Prefect Events client that records all events sent
-    to it for inspection during tests."""
-
-    events: List[Event]
-
-    def __init__(self):
-        self.events = []
-
-    async def emit(self, event: Event) -> None:
-        self.events.append(event)
-
-    async def __aenter__(self) -> "AssertingEventsClient":
-        self.events = []
-        return self
-
-    async def __aexit__(
-        self,
-        exc_type: Optional[Type[Exception]],
-        exc_val: Optional[Exception],
-        exc_tb: Optional[TracebackType],
-    ) -> "AssertingEventsClient":
-        self.events = []
-        return self
