+++ tests/events/test_clients.py
@@ -1,5 +1,6 @@
-from prefect.events.clients import AssertingEventsClient, NullEventsClient
+from prefect.events.clients import NullEventsClient
 from prefect.events.schemas import Event
+from prefect.testing.events import AssertingEventsClient
 
 
 async def test_null_client_can_emit_event(example_event: Event):
